[BMSPT-275] implementation of bcf zip creation with memory stream

+++ src/bcf-toolkit/Converter/Bcf21/Converter.cs
@@ -32,10 +32,20 @@ public class Converter : IConverter {
   ///   Defines the file writer function which must be used for write the BCF
   ///   object to the targeted version.
   /// </summary>
-  private readonly Dictionary<BcfVersionEnum, Func<IBcf, string, bool, Task<string>>> _writerFnMapper =
+  private readonly Dictionary<BcfVersionEnum, Func<IBcf, string, bool, Task<string>>> _writerToFolderFnMapper =
     new() {
-      [BcfVersionEnum.Bcf21] = FileWriter.WriteBcf,
-      [BcfVersionEnum.Bcf30] = Bcf30.FileWriter.WriteBcf
+      [BcfVersionEnum.Bcf21] = FileWriter.WriteBcfToFolder,
+      [BcfVersionEnum.Bcf30] = Bcf30.FileWriter.WriteBcfToFolder
+    };
+
+  /// <summary>
+  ///   Defines the file writer function which must be used for write the BCF
+  ///   object to the targeted version.
+  /// </summary>
+  private readonly Dictionary<BcfVersionEnum, Func<IBcf, Task<Stream>>> _writerToStreamFnMapper =
+    new() {
+      [BcfVersionEnum.Bcf21] = FileWriter.WriteBcfToStream,
+      [BcfVersionEnum.Bcf30] = Bcf30.FileWriter.WriteBcfToStream
     };
 
   public async Task BcfZipToJson(Stream source, string targetPath) {
@@ -69,29 +79,36 @@ public class Converter : IConverter {
       Version = new Version()
     };
 
-    await FileWriter.WriteBcf(bcf, target);
+    await FileWriter.WriteBcfToFolder(bcf, target);
   }
 
-  public async Task<Stream> ToBcfStream(IBcf bcf, BcfVersionEnum targetVersion) {
+  public async Task<Stream> ToBcfStream(IBcf bcf, BcfVersionEnum targetVersion, bool useTmpFolder) {
     var converterFn = _converterFnMapper[targetVersion];
     var convertedBcf = converterFn((Bcf)bcf);
 
     var workingDir = Directory.GetCurrentDirectory();
     var tmpBcfTargetPath = workingDir + $"/{Guid.NewGuid()}.bcfzip";
-    var writerFn = _writerFnMapper[targetVersion];
 
-    // keep the tmp files till the stream is created
-    var tmpFolder = await writerFn(convertedBcf, tmpBcfTargetPath, false);
-    var stream = new FileStream(tmpBcfTargetPath, FileMode.Open, FileAccess.Read);
+    // Using tmp folder or stream for bcf zip
+    if (useTmpFolder) {
+      var writerFn = _writerToFolderFnMapper[targetVersion];
+      // keep the tmp files till the stream is created
+      var tmpFolder = await writerFn(convertedBcf, tmpBcfTargetPath, false);
+      var stream = new FileStream(tmpBcfTargetPath, FileMode.Open, FileAccess.Read);
 
-    Directory.Delete(tmpFolder, true);
-    File.Delete(tmpBcfTargetPath);
+      Directory.Delete(tmpFolder, true);
+      File.Delete(tmpBcfTargetPath);
 
-    return stream;
+      return stream;
+    }
+    else {
+      var writerFn = _writerToStreamFnMapper[targetVersion];
+      return await writerFn(convertedBcf);
+    }
   }
 
   public Task ToBcfZip(IBcf bcf, string target) {
-    return FileWriter.WriteBcf((Bcf)bcf, target);
+    return FileWriter.WriteBcfToFolder((Bcf)bcf, target);
   }
 
   public Task ToJson(IBcf bcf, string target) {
